fix: address multiple code quality issues in audiocore
- Fix goroutine leak in birdnet analyzer by handling context cancellation
- Add metrics tracking for capture buffer write failures
- Fix division by zero in drop rate calculations
- Change NewProcessingPipeline to return error for better error handling
- Implement buffer pool usage for processed audio data efficiency
- Remove conditional sample rate mismatch allowance for stricter validation
- Replace runtime.SetFinalizer with runtime.AddCleanup (Go 1.24)
- Fix race condition in health monitor by computing silenceDuration on demand
- Add nil check in health monitor to prevent panic
- Fix integer overflow risk in chunk buffer calculations
- Use idiomatic time.Since() instead of now.Sub()
- Add goroutine leak detection in integration tests

+++ internal/audiocore/chunk_buffer_v2.go
@@ -30,8 +30,8 @@ type ChunkBufferV2 struct {
 
 // NewChunkBufferV2 creates a new improved chunk buffer
 func NewChunkBufferV2(config ChunkBufferConfig) *ChunkBufferV2 {
-	// Calculate target buffer size
-	bytesPerSecond := config.Format.SampleRate * config.Format.Channels * (config.Format.BitDepth / 8)
+	// Calculate target buffer size using int64 to prevent overflow
+	bytesPerSecond := int64(config.Format.SampleRate) * int64(config.Format.Channels) * int64(config.Format.BitDepth/8)
 	targetSize := int(float64(bytesPerSecond) * config.ChunkDuration.Seconds())
 
 	return &ChunkBufferV2{
